% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{check_gene_name_match_noDir}
\alias{check_gene_name_match_noDir}
\title{Function to Check if Gene IDs in the Input Gene Data Match Well When Direction Matrix Is Missing}
\usage{
check_gene_name_match_noDir(genes_in_Gene_sets, genes_in_Data)
}
\arguments{
\item{genes_in_Gene_sets}{A list of gene names from the gene sets.}

\item{genes_in_Data}{A list of gene names from the data matrix.}
}
\value{
If there are more than 10\\% of gene names in gene sets do not match those in the data matrix or direction matrix,
the single sample directional gene set analysis would stop.
}
\description{
This function is to check if the gene IDs in the gene sets and data matrix match well.
}
\details{
Before single sample directional gene set analysis, it is necessary to check whether the gene ID types in the in the
gene sets, data matrix and direction matrix have the same ID type. If not, users should double check to make gene ID types
match one another.
}
