% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_matrix_entrezID}
\alias{data_matrix_entrezID}
\title{This is data to be included in package}
\format{
An example data matrix of gene expressions with gene entrez ID as row names and
columns corresponding to different samples.
\describe{
\item{S1}{gene expression for this gene from the 1st sample}
\item{S2}{gene expression for this gene from the 2nd sample}
\item{S3}{gene expression for this gene from the 3rd sample}
\item{S4}{gene expression for this gene from the 4th sample}
\item{S5}{gene expression for this gene from the 5th sample}
\item{S6}{gene expression for this gene from the 6th sample}
\item{S7}{gene expression for this gene from the 7th sample}
\item{S8}{gene expression for this gene from the 8th sample}
\item{S9}{gene expression for this gene from the 9th sample}
\item{S10}{gene expression for this gene from the 10th sample}

}
}
\usage{
data_matrix_entrezID
}
\description{
This is data to be included in package
}
\keyword{datasets}
