% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{direction_matrix}
\alias{direction_matrix}
\title{This is data to be included in package}
\format{
An example direction matrix containing directionality information from summary statistics
such as effect size (ES) or p value, with each row for one gene.
\describe{
\item{gene}{gene entrez ID}
\item{ES}{effect size (SE) for this gene from summary statistics}
\item{pval}{p value for this gene from summary statistics}
}
}
\usage{
direction_matrix
}
\description{
This is data to be included in package
}
\keyword{datasets}
