% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_sets}
\alias{gene_sets}
\title{This is data to be included in package}
\format{
An example disease gene sets in the form of a list, with gene set names as list component names,
and each component is a vector of gene entrez ID. In this sample gene sets list, there are 10 gene sets in total.
\describe{
\item{TCELL.KEGG_T_CELL_RECEPTOR_SIGNALING_PATHWAY}{gene entrez ID related to this pathway}
\item{TLR.KEGG_TOLL_LIKE_RECEPTOR_SIGNALING_PATHWAY}{gene entrez ID related to this pathway}
\item{BCELL.KEGG_B_CELL_RECEPTOR_SIGNALING_PATHWAY}{gene entrez ID related to this pathway}
\item{NEUROTROPHIN.KEGG_NEUROTROPHIN_SIGNALING_PATHWAY}{gene entrez ID related to this pathway}
\item{ERBB.KEGG_ERBB_SIGNALING_PATHWAY}{gene entrez ID related to this pathway}
\item{CALCIUM.KEGG_CALCIUM_SIGNALING_PATHWAY}{gene entrez ID related to this pathway}
\item{CHEMOKINE.KEGG_CHEMOKINE_SIGNALING_PATHWAY}{gene entrez ID related to this pathway}
\item{GNRH.KEGG_GNRH_SIGNALING_PATHWAY}{gene entrez ID related to this pathway}
\item{VEGF.KEGG_VEGF_SIGNALING_PATHWAY}{gene entrez ID related to this pathway}
}
}
\usage{
gene_sets
}
\description{
This is data to be included in package
}
\keyword{datasets}
