% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{median_expression}
\alias{median_expression}
\title{Function to Calculate Median of Gene Expressions}
\usage{
median_expression(Data, pathway.db)
}
\arguments{
\item{Data}{Data matrix of gene expressions with gene ensembl ID as row names and columns corresponding to different samples.}

\item{pathway.db}{A list of gene sets.}
}
\value{
Matrix of average gene expression in each gene set with rows corresponding to gene sets and columns corresponding to
samples will be returned.
}
\description{
This function is to calculate the median of gene expressions for genes in the given gene sets.
}
\details{
Within the ssdGSA function, when GSA_method = "median.exprs", this function is used to calculate the average of
gene expressions for genes in the given gene sets.
}
